Huge update, pal info, bug fixes etc

+++ Assets/Jota/Stats.cs
@@ -5,7 +5,7 @@ public class Stats : ScriptableObject
 {
     public float health;
     public int attackDamage;
-    public int critChance;
+    public float critChance;
     public float critDamage = 1.75f;
     public int armor;
     public float alcoholAmplification;
